requirements.txt
asgiref==3.5.2
boto3==1.24.22
botocore==1.27.22
certifi==2022.6.15
Django==4.0.5
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==1.0.1
minio==7.1.9
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
pip~=22.0.2
attrs~=21.2.0
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
docutils~=0.17.1
keyring~=23.5.0
setuptools~=59.6.0
future~=0.18.2
requests~=2.25.1
pandas~=1.4.2